new update

+++ Backend/Schemas/CalenderEventModel.js
@@ -1,32 +1,33 @@
 module.exports = (sequelize, Sequelize) => {
-    const Calender = sequelize.define("calenderevent", {
-      id:{
+  const Calender = sequelize.define("calenderevent", {
+    id: {
       type: Sequelize.INTEGER,
       autoIncrement: true,
-      primaryKey: true},
-      title: {
-        type: Sequelize.STRING
-      },
-      description: {
-        type: Sequelize.STRING
-      },
-      enddate: {
-        type: Sequelize.STRING
-      },approved: {
-        type: Sequelize.TINYINT
-      },
+      primaryKey: true
+    },
+    title: {
+      type: Sequelize.STRING
+    },
+    description: {
+      type: Sequelize.STRING
+    },
+    enddate: {
+      type: Sequelize.STRING
+    }, approved: {
+      type: Sequelize.SMALLINT
+    },
 
-      start: {
-        type: Sequelize.STRING
-      },
-      event: {
-        type: Sequelize.TINYINT
-      },
-      type: {
-        type: Sequelize.TINYINT
-      },
-      
-    });
-    return Calender;
-  };
+    start: {
+      type: Sequelize.STRING
+    },
+    event: {
+      type: Sequelize.SMALLINT
+    },
+    type: {
+      type: Sequelize.SMALLINT
+    },
+
+  });
+  return Calender;
+};
