+++ Backend/Schemas/FacilityModel.js
@@ -1,25 +1,27 @@
 module.exports = (sequelize, Sequelize) => {
-    const Post = sequelize.define("tblfacility", {
-      title: {
-        type: Sequelize.STRING
-      },
-      description: {
-        type: Sequelize.STRING
-      },
-      imagelink: {
-        type: Sequelize.STRING
-      },approved: {
-        type: Sequelize.TINYINT
-      },
-      id:{
-        type: Sequelize.INTEGER,
-        autoIncrement: true,
-        primaryKey: true},
-      public_id: {
-        type: Sequelize.STRING
-      },
-      status: {
-        type: Sequelize.TINYINT
-      },  });
-    return Post;
-  };
+  const Post = sequelize.define("tblfacility", {
+    title: {
+      type: Sequelize.STRING
+    },
+    description: {
+      type: Sequelize.STRING
+    },
+    imagelink: {
+      type: Sequelize.STRING
+    }, approved: {
+      type: Sequelize.SMALLINT
+    },
+    id: {
+      type: Sequelize.INTEGER,
+      autoIncrement: true,
+      primaryKey: true
+    },
+    public_id: {
+      type: Sequelize.STRING
+    },
+    status: {
+      type: Sequelize.SMALLINT
+    },
+  });
+  return Post;
+};
