+++ Backend/Schemas/VacancyModel.js
@@ -1,26 +1,27 @@
 module.exports = (sequelize, Sequelize) => {
-    const Vacancy = sequelize.define("tblvacancy", {
-      id:{
+  const Vacancy = sequelize.define("tblvacancy", {
+    id: {
       type: Sequelize.INTEGER,
       autoIncrement: true,
-      primaryKey: true},
-      title: {
-        type: Sequelize.STRING
-      },
-      description: {
-        type: Sequelize.STRING
-      },
-      fromdate:{
-        type: Sequelize.STRING
-      },
-      todate:{
-        type: Sequelize.STRING
-      },
-      status: {
-        type: Sequelize.TINYINT
-      },
-     
-      
-    });
-    return Vacancy;
-  };
+      primaryKey: true
+    },
+    title: {
+      type: Sequelize.STRING
+    },
+    description: {
+      type: Sequelize.STRING
+    },
+    fromdate: {
+      type: Sequelize.STRING
+    },
+    todate: {
+      type: Sequelize.STRING
+    },
+    status: {
+      type: Sequelize.SMALLINT
+    },
+
+
+  });
+  return Vacancy;
+};
